refactor: define `embedAttributesInTags` as generic HTML test helper
Refactor the `embedAttributesInTags` from both the CSS classes and
variables test suite into the html-helpers.ts module (previously html-
constants.ts) so that it can be reused accross test suites.
