+++ packages/core/src/manglers/__tests__/css-variables.test.ts
@@ -9,7 +9,11 @@ import {
   CSS_VALUES,
   CSS_VALUES_NO_STRINGS,
 } from "./css-constants";
-import { SELF_CLOSING_TAGS, STANDARD_TAGS } from "./html-constants";
+import {
+  embedAttributesInTags,
+  SELF_CLOSING_TAGS,
+  STANDARD_TAGS,
+} from "./html-helpers";
 import {
   getArrayOfFormattedStrings,
   varyQuotes,
@@ -308,21 +312,6 @@ suite("CSS Variable Mangler", function() {
       },
     ];
 
-    const embedAttributesInTags = (testCase: TestCase): TestCase[] => {
-      return [
-        ...STANDARD_TAGS
-          .map((tag: string): TestCase => ({
-            input: `<${tag} ${testCase.input}>Lorem ipsum</${tag}>`,
-            expected: `<${tag} ${testCase.expected}>Lorem ipsum</${tag}>`,
-          })),
-        ...SELF_CLOSING_TAGS
-          .map((tag: string): TestCase => ({
-            input: `<${tag} ${testCase.input}/>`,
-            expected: `<${tag} ${testCase.expected}/>`,
-          }))
-          .flatMap((_testCase) => varySpacing("/", _testCase)),
-      ];
-    };
     const embedDeclarationsInAttributes = (testCase: TestCase): TestCase[] => {
       return [
         {
@@ -362,17 +351,17 @@ suite("CSS Variable Mangler", function() {
           .flatMap(embedAttributesInTags),
       },
       {
-        name: "varying spacing in declarations",
-        cases: SAMPLE_VARIABLE_DECLARATIONS
-          .flatMap(embedDeclarationsInAttributes)
-          .flatMap((testCase) => varySpacing([":", "(", ",", ")", ";"], testCase))
-          .flatMap(embedAttributesInTags),
-      },
-      {
         name: "varying quotes",
         cases: SAMPLE_VARIABLE_DECLARATIONS
           .flatMap(embedDeclarationsInAttributes)
           .flatMap((testCase) => varyQuotes("html", testCase))
+          .flatMap(embedAttributesInTags),
+      },
+      {
+        name: "varying spacing in declarations",
+        cases: SAMPLE_VARIABLE_DECLARATIONS
+          .flatMap(embedDeclarationsInAttributes)
+          .flatMap((testCase) => varySpacing([":", "(", ",", ")", ";"], testCase))
           .flatMap(embedAttributesInTags),
       },
       {
