--- packages/core/src/manglers/__tests__/html-constants.ts
@@ -1,16 +1,0 @@
-/**
- * A list of standard (non self-closing) HTML tags.
- */
-export const STANDARD_TAGS: string[] = [
-  "div",
-  "header",
-  "p",
-];
-
-/**
- * A list of standard self-closing HTML tags.
- */
-export const SELF_CLOSING_TAGS: string[] = [
-  "img",
-  "input",
-];
